requirements.txt
numpy==1.16.4
pandas==1.0.3
scikit-learn==0.22.2
azureml_train_restclients_hyperdrive==1.5.0
azureml-sdk==1.5.0
pytest==5.4.0
pytest-cov==2.8.0
requests==2.23.0
python-dotenv==0.12.0
flake8==3.7.0
flake8_formatter_junit_xml==0.0.6
azure-cli==2.5
